[gym/common] Move 'sample' from 'spaces.py' to 'helpers.py'.
