+++ python/gym_jiminy/common/gym_jiminy/common/utils/helpers.py
@@ -2,7 +2,8 @@
 """
 import math
 import logging
-from typing import List, Sequence, Union, ValuesView
+from functools import partial
+from typing import Any, List, Sequence, ValuesView, Optional, Union, Protocol
 
 import gymnasium as gym
 import numpy as np
@@ -11,14 +12,25 @@ import numba as nb
 import jiminy_py.core as jiminy
 from jiminy_py import tree
 
-from .spaces import FieldNested, DataNested, zeros
+from .spaces import FieldNested, DataNested, ArrayOrScalar, zeros
 
 
 LOGGER = logging.getLogger(__name__)
 
+GLOBAL_RNG = np.random.default_rng()
+
 
 FieldNestedSequence = Sequence[Union['FieldNestedSequence', str]]
 FieldNestedList = List[Union[FieldNestedSequence, str]]
+
+
+class RandomDistribution(Protocol):
+    """Protocol to be satisfied when passing generic statistical distribution
+    callable to `sample` method.
+    """
+    def __call__(self, rg: np.random.Generator, *args: Any, **kwargs: Any
+                 ) -> ArrayOrScalar:
+        ...
 
 
 @nb.jit(nopython=True, cache=True, inline='always')
@@ -119,3 +131,89 @@ def register_variables(controller: jiminy.AbstractController,
         assert isinstance(fieldname, list), (
             f"'fieldname' ({fieldname}) should be a list of strings.")
         controller.register_variables(fieldname, value)
+
+
+def sample(low: Union[float, np.ndarray] = -1.0,
+           high: Union[float, np.ndarray] = 1.0,
+           dist: Union[str, RandomDistribution] = 'uniform',
+           scale: Union[float, np.ndarray] = 1.0,
+           enable_log_scale: bool = False,
+           shape: Optional[Sequence[int]] = None,
+           rg: Optional[np.random.Generator] = None
+           ) -> np.ndarray:
+    """Randomly sample values from a given distribution.
+
+    .. note:
+        If 'low', 'high', and 'scale' are floats, then the output is float if
+        'shape' is None, otherwise it has type `np.ndarray` and shape 'shape'.
+        Similarly, if any of 'low', 'high', and 'scale' are `np.ndarray`, then
+        its shape follows the broadcasting rules between these variables.
+
+    :param low: Lower value for bounded distribution, negative-side standard
+                deviation otherwise.
+                Optional: -1.0 by default.
+    :param high: Upper value for bounded distribution, positive-side standard
+                 deviation otherwise.
+                 Optional: 1.0 by default.
+    :param dist: The statistical from which to draw samples, either provided as
+                 a pre-defined string or a callable. For strings, then it must
+                 be a member function of `np.random.Generator` (only 'uniform'
+                 and 'normal' are supported for now). For callables, it must
+                 corresponds to a standardized distribution and satisfying
+                 `gym_jiminy.common.utils.RandomDistribution` protocol. This is
+                 especially useful for specifying custom parameters of complex
+                 distributions such as Beta. Using `functools.partial` is
+                 recommended, eg `partial(np.random.Generator.Beta, a=1, b=8)`.
+                 Optional: 'uniform' by default.
+    :param scale: Shrink the standard deviation of the distribution around the
+                  mean by this factor.
+                  Optional: No scaling by default?
+    :param enable_log_scale: The sampled values are power of 10.
+    :param shape: Enforce of the sampling shape. Only available if 'low',
+                  'high' and 'scale' are floats. `None` to disable.
+                  Optional: Disabled by default.
+    :param rg: Custom random number generator from which to draw samples.
+               Optional: Default to `np.random`.
+    """
+    # Compute mean and deviation from low and high arguments
+    mean = 0.5 * (low + high)
+    dev = 0.5 * scale * (high - low)
+
+    # Get sample shape.
+    # Better use dev than mean since it works even if only scale is array.
+    if isinstance(dev, np.ndarray):
+        if shape is None:
+            shape = dev.shape
+        else:
+            try:
+                shape = list(shape)
+                np.broadcast(np.empty(shape, dtype=[]), dev)
+            except ValueError as e:
+                raise ValueError(
+                    f"'shape' {shape} must be broadcast-able with 'low', "
+                    f"'high' and 'scale' {dev.shape} if specified.") from e
+
+    # Define "standardized" distribution callable if only its name was provided
+    if isinstance(dist, str):
+        if dist not in ('uniform', 'normal'):
+            raise NotImplementedError(
+                f"'{dist}' distribution type is not supported for now.")
+        dist_fn = getattr(np.random.Generator, dist)
+        if dist == 'uniform':
+            # The uniform distribution is NOT standardized by default
+            dist_fn = partial(dist_fn, low=-1.0, high=1.0)
+    else:
+        dist_fn = dist
+
+    # Generate samples from distribution.
+    # Make sure that the result is always returned as np.ndarray.
+    value = np.asarray(dist_fn(rg or GLOBAL_RNG, size=shape))
+
+    # Apply mean and standard deviation transformation
+    value = mean + dev * value
+
+    # Revert log scale if requested
+    if enable_log_scale:
+        value = 10 ** value
+
+    return value
